fix-report-bug

+++ takin-web-data/src/main/java/com/pamirs/takin/entity/domain/dto/report/ReportCostDTO.java
@@ -7,5 +7,5 @@ import java.io.Serializable;
 @Data
 public class ReportCostDTO implements Serializable {
 
-    public Integer count;
+    public Long count;
 }
